requirements.txt
fastapi~=0.115.6
SQLAlchemy~=2.0.37
DateTime~=5.5
PyJWT~=2.10.1
passlib~=1.7.4
python-dotenv~=1.0.1